requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.6.0
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.0.7
django-debug-toolbar==4.4.6
django-debug-toolbar-force==0.2
django-filter==24.2
django-nine==0.2.7
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
drf-social-oauth2==3.0.1
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
